+++ sda/cmd/mapper/mapper.go
@@ -54,8 +54,8 @@ func run() error {
 		return fmt.Errorf("failed to initialize sda db, due to: %v", err)
 	}
 	defer db.Close()
-	if dbSchemaVersion, err := db.SchemaVersion(); err != nil || dbSchemaVersion < 23 {
-		return errors.Join(errors.New("database schema v23 is required"), err)
+	if dbSchemaVersion, err := db.SchemaVersion(); err != nil || dbSchemaVersion < 25 {
+		return errors.Join(errors.New("database schema v25 is required"), err)
 	}
 
 	mqBroker, err = broker.NewMQ(conf.Broker)
@@ -174,7 +174,15 @@ func handleMessage(ctx context.Context, delivered amqp.Delivery) {
 	case "mapping":
 		log.Debug("mapping type operation, mapping files to dataset")
 		for _, aID := range mappings.AccessionIDs {
-			log.Debugf("Mapped file to dataset (correlation-id: %s, dataset-id: %s, accession-id: %s)", delivered.CorrelationId, mappings.DatasetID, aID)
+			var fileDownloadPath *string
+
+			if mappings.FileDownloadPaths != nil {
+				if v, ok := mappings.FileDownloadPaths[aID]; ok && v != "" {
+					fileDownloadPath = &v
+				}
+			}
+
+			log.Debugf("Mapping file to dataset (correlation-id: %s, dataset-id: %s, accession-id: %s, file-download-path overridden: %t)", delivered.CorrelationId, mappings.DatasetID, aID, fileDownloadPath != nil)
 			fileMappingData, err := tx.GetMappingData(ctx, aID)
 			if err != nil {
 				log.Errorf("failed to get file info for file with accession-id: %s, can not map file to dataset: %s, due to: %v", aID, mappings.DatasetID, err)
@@ -195,7 +203,7 @@ func handleMessage(ctx context.Context, delivered amqp.Delivery) {
 
 				return
 			}
-			if err := tx.MapFileToDataset(ctx, mappings.DatasetID, fileMappingData.FileID); err != nil {
+			if err := tx.MapFileToDataset(ctx, mappings.DatasetID, fileMappingData.FileID, fileDownloadPath); err != nil {
 				log.Errorf("failed to map file: %s to dataset-id: %s, reason: %v", fileMappingData.FileID, mappings.DatasetID, err)
 
 				// Nack message so the server gets notified that something is wrong and requeue the message